+++ src/test/java/org/realityforge/replicant/server/transport/PacketTest.java
@@ -40,5 +40,6 @@ public class PacketTest
     assertFalse( packet.isPrevious( 0 ) );
     assertFalse( packet.isPrevious( 2 ) );
 
+    assertFalse( packet.isLessThanOrEqual( Integer.MIN_VALUE ) );
   }
 }